rearanged structure of web app files
moved app.py into __init__.py and moved app.py outside of the application folder to server as a pseudo main.py to follow a tutorial
