+++ Flask_Application/README.md
@@ -1,2 +1,28 @@
 # Senior-Project
 Senior project for 2024 (Will add what project later)
+# Idea management web application
+
+## Setup & Installation
+
+Make sure you have the latest version of Python installed.
+
+```bash
+git clone <repo-url>
+```
+# run this to install requried packages
+```bash
+pip install -r requirements.txt
+```
+# To run the web app do run this in the terminal
+```bash
+python app.py
+```
+## Running The App **WHEN FINISHED**
+
+```bash
+python main.py
+```
+
+## Viewing The App
+
+Go to `http://127.0.0.1:5000`
